lmdb_native.dart: use Isolate.resolvePackageUriSync() to resolve URI
If we install the package from anywhere, we need to know the
absolute path of the native library file/directory. There are several
proposals to solve this problem for Dart, but none of them seems to be
universally applicable on all platforms.

Use Isolate.resolvePackageUriSync(), which seems to be the only way of
resolving the absolute path at least on non-iOS platforms.

On MacOS, when using a test project to load dart_lmdb2, the following
path is resolved:

 ${PRJ_DIR}/dart_lmdb2/lib/src/native/liblmdb.dylib

Which is correct.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>